Module stm32wlxx_hal::adc

Analog to digital converter

Quickstart:

Adc::pin Sample an analog pin

Adc::temperature Sample the junction temperature

Adc::vbat Sample the battery voltage

Modules

irq

Interrupt masks

Structs

AdcNon-stm32wl5x_cm0p

Analog to digital converter driver.

Enums

Ch

ADC channels

Clk

ADC clock mode

OversampleRatioNon-stm32wl5x_cm0p

OVSR0

OversampleShiftNon-stm32wl5x_cm0p

OVSS0

Ts

ADC sample times

Constants

TS_MIN_SAMPLE

tS_temp temperature sensor minimum sampling time

TS_START_MAX

tSTART temperature sensor maximum startup time when entering continuous mode

TS_START_TYP

tSTART temperature sensor typical startup time when entering continuous mode

T_ADCVREG_SETUP

tADCVREG_SETUP ADC voltage regulator maximum startup time

T_ADCVREG_SETUP_MICROS

T_ADCVREG_SETUP expressed in microseconds

Functions

vref_cal

Internal voltage reference ADC calibration