Struct stm32wlxx_hal::flash::Page

pub struct Page { /* private fields */ }

Page address.

Implementations

impl Page

pub const SIZE: usize = 2_048usize

Page size in bytes.

pub const unsafe fn from_index_unchecked(idx: u8) -> Self

Create a page address from an index without checking bounds.

Safety

The idx argument must be a valid page number, less than the value returned by num_pages.

Example

use stm32wlxx_hal::flash::Page;

let page0 = unsafe { Page::from_index_unchecked(0) };

pub fn from_index(idx: u8) -> Option<Self>

Create a page address from an index.

Returns None if the value index is greater than the index of the last page, for example 0x7F (page 127) on the STM32WLE5.

Example

use stm32wlxx_hal::flash::Page;

assert!(Page::from_index(8).is_some());
assert!(Page::from_index(128).is_none());

pub fn from_byte_offset(offset: usize) -> Option<Self>

Create a page address from an offset from the base of the flash memory.

Returns None if the address is out of bounds, or not page aligned.

Example

use stm32wlxx_hal::flash::Page;

assert_eq!(Page::from_byte_offset(0), Page::from_index(0));
assert_eq!(Page::from_byte_offset(2048), Page::from_index(1));
assert!(Page::from_byte_offset(2047).is_none());
assert!(Page::from_byte_offset(usize::MAX).is_none());

pub fn from_addr(addr: usize) -> Option<Self>

Create a page address from an absolute address.

Returns None if the address is out of bounds, or not page aligned.

Example

use stm32wlxx_hal::flash::Page;

assert_eq!(Page::from_addr(0x0800_0000), Page::from_index(0));
assert_eq!(Page::from_addr(0x0800_0800), Page::from_index(1));
assert!(Page::from_addr(0).is_none());
assert!(Page::from_addr(usize::MAX).is_none());
assert!(Page::from_addr(0x0800_0001).is_none());

pub const fn to_index(self) -> u8

Get the page index.

Example

use stm32wlxx_hal::flash::Page;

let page7 = unsafe { Page::from_index_unchecked(7) };
assert_eq!(page7.to_index(), 7);

pub const fn addr(&self) -> usize

Get the page address.

Example

use stm32wlxx_hal::flash::Page;

let page127 = unsafe { Page::from_index_unchecked(127) };
let page0 = unsafe { Page::from_index_unchecked(0) };

assert_eq!(page0.addr(), 0x0800_0000);
assert_eq!(page127.addr(), 0x0803_F800);

pub const fn addr_range(&self) -> Range<usize>

Get the address range of the page.

Example

use core::ops::Range;
use stm32wlxx_hal::flash::Page;

let page0 = unsafe { Page::from_index_unchecked(0) };
assert_eq!(
    page0.addr_range(),
    Range {
        start: 0x0800_0000,
        end: 0x0800_0800
    }
);