Struct stm32wlxx_hal::gpio::PortA

pub struct PortA {

    pub a0: A0,
    pub a1: A1,
    pub a2: A2,
    pub a3: A3,
    pub a4: A4,
    pub a5: A5,
    pub a6: A6,
    pub a7: A7,
    pub a8: A8,
    pub a9: A9,
    pub a10: A10,
    pub a11: A11,
    pub a12: A12,
    pub a13: A13,
    pub a14: A14,
    pub a15: A15,

}

Port A GPIOs

Fields
a0: A0a1: A1a2: A2a3: A3a4: A4a5: A5a6: A6a7: A7a8: A8a9: A9a10: A10a11: A11a12: A12a13: A13a14: A14a15: A15
Implementations

impl PortA

pub fn split(gpioa: GPIOA, rcc: &mut RCC) -> Self

Reset GPIO port A and split the port into individual pins.

This will enable clocks and reset the GPIO port.

Example

Get GPIO A0.

use stm32wlxx_hal::{
    gpio::{pins, PortA},
    pac,
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

let gpioa: PortA = PortA::split(dp.GPIOA, &mut dp.RCC);
let a0: pins::A0 = gpioa.a0;

pub unsafe fn steal() -> Self

Steal the port A GPIOs from whatever is currently using them.

This will not initialize the GPIOs (unlike split).

Safety

This will create new GPIOs, bypassing the singleton checks that normally occur. You are responsible for ensuring that the driver has exclusive access to the GPIOs. You are also responsible for ensuring the GPIO peripheral has been setup correctly.

Example

use stm32wlxx_hal::gpio::PortA;

// ... setup happens here

let gpioa: PortA = unsafe { PortA::steal() };

pub unsafe fn disable_clock(rcc: &mut RCC)

Disable the GPIOA clock.

Safety

You cannot use any port-A GPIO pin while the clock is disabled.

You are responsible for re-enabling the clock before resuming use of any port A GPIO.

pub fn enable_clock(rcc: &mut RCC)

Enable the GPIOA clock.

pub fn set_sleep_clock(rcc: &mut RCC, enable: bool)

Set the IO port clock enable during sleep mode.

false: Clock disabled during Sleep and Stop modes

true: Clock enabled during Sleep mode, disabled during Stop mode

pub fn input_level() -> u16

Get the input level of all pins on this port.

pub fn output_level() -> u16

Get the output level of all pins on this port.