Struct stm32wlxx_hal::gpio::PortB

pub struct PortB {

    pub b0: B0,
    pub b1: B1,
    pub b2: B2,
    pub b3: B3,
    pub b4: B4,
    pub b5: B5,
    pub b6: B6,
    pub b7: B7,
    pub b8: B8,
    pub b9: B9,
    pub b10: B10,
    pub b11: B11,
    pub b12: B12,
    pub b13: B13,
    pub b14: B14,
    pub b15: B15,

}

Port B GPIOs

Fields
b0: B0b1: B1b2: B2b3: B3b4: B4b5: B5b6: B6b7: B7b8: B8b9: B9b10: B10b11: B11b12: B12b13: B13b14: B14b15: B15
Implementations

impl PortB

pub fn split(gpiob: GPIOB, rcc: &mut RCC) -> Self

Reset GPIO port B and split the port into individual pins.

This will enable clocks and reset the GPIO port.

Example

Get GPIO B0.

use stm32wlxx_hal::{
    gpio::{pins, PortB},
    pac,
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let b0: pins::B0 = gpiob.b0;

pub unsafe fn steal() -> Self

Steal the port B GPIOs from whatever is currently using them.

This will not initialize the GPIOs (unlike split).

Safety

This will create new GPIOs, bypassing the singleton checks that normally occur. You are responsible for ensuring that the driver has exclusive access to the GPIOs. You are also responsible for ensuring the GPIO peripheral has been setup correctly.

Example

use stm32wlxx_hal::gpio::PortB;

// ... setup happens here

let gpioa: PortB = unsafe { PortB::steal() };

pub unsafe fn disable_clock(rcc: &mut RCC)

Disable the GPIOB clock.

Safety

You cannot use any port-B GPIO pin while the clock is disabled.

You are responsible for re-enabling the clock before resuming use of any port B GPIO.

pub fn enable_clock(rcc: &mut RCC)

Enable the GPIOB clock.

pub fn set_sleep_clock(rcc: &mut RCC, enable: bool)

Set the IO port clock enable during sleep mode.

false: Clock disabled during Sleep and Stop modes

true: Clock enabled during Sleep mode, disabled during Stop mode

pub fn input_level() -> u16

Get the input level of all pins on this port.

pub fn output_level() -> u16

Get the output level of all pins on this port.