Struct stm32wlxx_hal::gpio::PortC

pub struct PortC {
    pub c0: C0,
    pub c1: C1,
    pub c2: C2,
    pub c3: C3,
    pub c4: C4,
    pub c5: C5,
    pub c6: C6,
    pub c13: C13,
    pub c14: C14,
    pub c15: C15,
}

Port C GPIOs

Fields
c0: C0c1: C1c2: C2c3: C3c4: C4c5: C5c6: C6c13: C13c14: C14c15: C15
Implementations

impl PortC

pub fn split(gpioc: GPIOC, rcc: &mut RCC) -> Self

Reset GPIO port C and split the port into individual pins.

This will enable clocks and reset the GPIO port.

Example

Get GPIO C0.

use stm32wlxx_hal::{
    gpio::{pins, PortC},
    pac,
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

let gpioc: PortC = PortC::split(dp.GPIOC, &mut dp.RCC);
let c0: pins::C0 = gpioc.c0;

pub unsafe fn steal() -> Self

Steal the port C GPIOs from whatever is currently using them.

This will not initialize the GPIOs (unlike split).

Safety

This will create new GPIOs, bypassing the singleton checks that normally occur. You are responsible for ensuring that the driver has exclusive access to the GPIOs. You are also responsible for ensuring the GPIO peripheral has been setup correctly.

Example

use stm32wlxx_hal::gpio::PortC;

// ... setup happens here

let gpioa: PortC = unsafe { PortC::steal() };

pub unsafe fn disable_clock(rcc: &mut RCC)

Disable the GPIOC clock.

Safety

You cannot use any port-C GPIO pin while the clock is disabled.

You are responsible for re-enabling the clock before resuming use of any port C GPIO.

pub fn enable_clock(rcc: &mut RCC)

Enable the GPIOC clock.

pub fn input_level() -> u16

Get the input level of all pins on this port.

pub fn output_level() -> u16

Get the output level of all pins on this port.