Struct stm32wlxx_hal::lptim::LpTim3

pub struct LpTim3 { /* private fields */ }

Low-power timer 3 driver.

Implementations

impl LpTim3

pub fn new_trigger_pin(
&mut self,
pin: A11,
filter: Filter,
pol: TrgPol,
cs: &CriticalSection
) -> LpTim3Trg

Setup a new pin trigger.

Panics

(debug) timer is enabled.

Example

Setup a one-shot timer that starts after a transition on pin A11.

use stm32wlxx_hal::{
    cortex_m,
    embedded_hal::timer::CountDown,
    gpio::PortA,
    lptim::{self, Filter, LpTim, LpTim3, LpTim3Trg, Prescaler::Div1, TrgPol},
    pac,
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let pa: PortA = PortA::split(dp.GPIOA, &mut dp.RCC);
let mut lptim3: LpTim3 = LpTim3::new(dp.LPTIM3, lptim::Clk::Hsi16, Div1, &mut dp.RCC);
let lptim3trg: LpTim3Trg = cortex_m::interrupt::free(|cs| {
    lptim3.new_trigger_pin(pa.a11, Filter::Any, TrgPol::Both, cs)
});
lptim3.start(12_345_u16);
// timer will only start after any transition on pin A11
nb::block!(lptim3.wait());

pub fn free_trigger_pin(&mut self, pin: LpTim3Trg) -> A11

Free the trigger pin previously created with new_trigger_pin.

This is will the trigger source to a software trigger.

Panics

(debug) timer is enabled.

Example

Setup a one-shot timer that starts after a transition on pin A11.

use stm32wlxx_hal::{
    embedded_hal::timer::CountDown,
    gpio::PortA,
    lptim::{self, Filter, LpTim, LpTim3, LpTim3Trg, Prescaler::Div1, TrgPol},
    pac,
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let pa: PortA = PortA::split(dp.GPIOA, &mut dp.RCC);
let mut lptim3: LpTim3 = LpTim3::new(dp.LPTIM3, lptim::Clk::Hsi16, Div1, &mut dp.RCC);
let lptim3trg: LpTim3Trg = cortex_m::interrupt::free(|cs| {
    lptim3.new_trigger_pin(pa.a11, Filter::Any, TrgPol::Both, cs)
});
lptim3.start(12_345_u16);
// timer will only start after any transition on pin A11
nb::block!(lptim3.wait());

// free the pin
let a11 = lptim3.free_trigger_pin(lptim3trg);

impl LpTim3

pub fn new_output_pin(&mut self, pin: A1, cs: &CriticalSection) -> LpTim3OutPin

Setup a new output pin.

pub fn free_output_pin(&mut self, pin: LpTim3OutPin) -> A1

Free the output pin previously created with new_output_pin.