Module stm32wlxx_hal::pac

Available on crate feature stm32wl5x_cm4 only.

Peripheral access crate Peripheral access API for STM32WL5X_CM4 microcontrollers (generated using svd2rust v0.24.1 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

adc

Analog to digital convertor

aes

Advanced encryption standard hardware accelerator 1

comp

Comparator

crc

Cyclic redundancy check calculation unit

dac

Digital-to-analog converter

dbgmcu

Microcontroller Debug Unit

dma1

Direct memory access controller

dmamux

DMA request multiplexer

exti

External interrupt/event controller

flash

Flash

gpioa

General-purpose I/Os

gpiob

General-purpose I/Os

gpioc

General-purpose I/Os

gpioh

General-purpose I/Os

hsem

Hardware semaphore

i2c1

Inter-integrated circuit

ipcc

Inter Processor communication controller

iwdg

Independent watchdog

lptim1

Low-power timer

lptim2

Low-power timer

lptim3

Low-power timer

lpuart

Universal synchronous asynchronous receiver transmitter

nvic_stir

Nested vectored interrupt controller

pka

Public key accelerator

pwr

Power control

rcc

Reset and clock control

rng

True random number generator

rtc

Real-time clock

scb_actrl

System control block ACTLR

spi1

Serial peripheral interface/Inter-IC sound

stk

SysTick timer

syscfg

System configuration controller

tamp

Tamper and backup registers

tim1

Advanced-control timers

tim2

General-purpose-timers

tim16

General-purpose timers

tim17

General-purpose timers

tzic

TrustZone Interrupt Control

tzsc

Global TrustZone Controller

usart1

Universal synchronous asynchronous receiver transmitter

vrefbuf

Voltage reference buffer

wwdg

System window watchdog

Structs

ADC

Analog to digital convertor

AES

Advanced encryption standard hardware accelerator 1

CBP

Cache and branch predictor maintenance operations

COMP

Comparator

CPUID

CPUID

CRC

Cyclic redundancy check calculation unit

CorePeripherals

Core peripherals

DAC

Digital-to-analog converter

DBGMCU

Microcontroller Debug Unit

DCB

Debug Control Block

DMA1

Direct memory access controller

DMA2

Direct memory access controller

DMAMUX

DMA request multiplexer

DWT

Data Watchpoint and Trace unit

EXTI

External interrupt/event controller

FLASH

Flash

FPB

Flash Patch and Breakpoint unit

FPU

Floating Point Unit

GPIOA

General-purpose I/Os

GPIOB

General-purpose I/Os

GPIOC

General-purpose I/Os

GPIOH

General-purpose I/Os

HSEM

Hardware semaphore

I2C1

Inter-integrated circuit

I2C2

Inter-integrated circuit

I2C3

Inter-integrated circuit

IPCC

Inter Processor communication controller

ITM

Instrumentation Trace Macrocell

IWDG

Independent watchdog

LPTIM1

Low-power timer

LPTIM2

Low-power timer

LPTIM3

Low-power timer

LPUART

Universal synchronous asynchronous receiver transmitter

MPU

Memory Protection Unit

NVIC

Nested Vector Interrupt Controller

NVIC_STIR

Nested vectored interrupt controller

PKA

Public key accelerator

PWR

Power control

Peripherals

All the peripherals

RCC

Reset and clock control

RNG

True random number generator

RTC

Real-time clock

SCB

System Control Block

SCB_ACTRL

System control block ACTLR

SPI1

Serial peripheral interface/Inter-IC sound

SPI2

Serial peripheral interface/Inter-IC sound

SPI3

Serial peripheral interface/Inter-IC sound

STK

SysTick timer

SYSCFG

System configuration controller

SYST

SysTick: System Timer

TAMP

Tamper and backup registers

TIM1

Advanced-control timers

TIM2

General-purpose-timers

TIM16

General-purpose timers

TIM17

General-purpose timers

TPIU

Trace Port Interface Unit

TZIC

TrustZone Interrupt Control

TZSC

Global TrustZone Controller

USART1

Universal synchronous asynchronous receiver transmitter

USART2

Universal synchronous asynchronous receiver transmitter

VREFBUF

Voltage reference buffer

WWDG

System window watchdog

Enums

Interrupt

Enumeration of all the interrupts.

interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority

Attribute Macros

interrupt