Module stm32wlxx_hal::pac::aes::cr

Available on crate feature stm32wl5x_cm4 only.

control register

Structs

CR_SPEC

control register

R

Register CR reader

W

Register CR writer

Enums

CCFC_A

Computation Complete Flag Clear

CCFIE_A

CCF flag interrupt enable

CHMOD2_A

AES chaining mode Bit2

CHMOD_A

AES chaining mode Bit1 Bit0

DATATYPE_A

Data type selection (for data in and data out to/from the cryptographic block)

DMAINEN_A

Enable DMA management of data input phase

DMAOUTEN_A

Enable DMA management of data output phase

EN_A

AES enable

ERRC_A

Error clear

ERRIE_A

Error interrupt enable

GCMPH_A

Used only for GCM, CCM and GMAC algorithms and has no effect when other algorithms are selected

KEYSIZE_A

Key size selection

MODE_A

AES operating mode

Type Definitions

CCFC_R

Field CCFC reader - Computation Complete Flag Clear

CCFC_W

Field CCFC writer - Computation Complete Flag Clear

CCFIE_R

Field CCFIE reader - CCF flag interrupt enable

CCFIE_W

Field CCFIE writer - CCF flag interrupt enable

CHMOD2_R

Field CHMOD2 reader - AES chaining mode Bit2

CHMOD2_W

Field CHMOD2 writer - AES chaining mode Bit2

CHMOD_R

Field CHMOD reader - AES chaining mode Bit1 Bit0

CHMOD_W

Field CHMOD writer - AES chaining mode Bit1 Bit0

DATATYPE_R

Field DATATYPE reader - Data type selection (for data in and data out to/from the cryptographic block)

DATATYPE_W

Field DATATYPE writer - Data type selection (for data in and data out to/from the cryptographic block)

DMAINEN_R

Field DMAINEN reader - Enable DMA management of data input phase

DMAINEN_W

Field DMAINEN writer - Enable DMA management of data input phase

DMAOUTEN_R

Field DMAOUTEN reader - Enable DMA management of data output phase

DMAOUTEN_W

Field DMAOUTEN writer - Enable DMA management of data output phase

EN_R

Field EN reader - AES enable

EN_W

Field EN writer - AES enable

ERRC_R

Field ERRC reader - Error clear

ERRC_W

Field ERRC writer - Error clear

ERRIE_R

Field ERRIE reader - Error interrupt enable

ERRIE_W

Field ERRIE writer - Error interrupt enable

GCMPH_R

Field GCMPH reader - Used only for GCM, CCM and GMAC algorithms and has no effect when other algorithms are selected

GCMPH_W

Field GCMPH writer - Used only for GCM, CCM and GMAC algorithms and has no effect when other algorithms are selected

KEYSIZE_R

Field KEYSIZE reader - Key size selection

KEYSIZE_W

Field KEYSIZE writer - Key size selection

MODE_R

Field MODE reader - AES operating mode

MODE_W

Field MODE writer - AES operating mode

NPBLB_R

Field NPBLB reader - Number of padding bytes in last block of payload

NPBLB_W

Field NPBLB writer - Number of padding bytes in last block of payload