Module stm32wlxx_hal::pac::exti::rtsr2

Available on crate feature stm32wl5x_cm4 only.

rising trigger selection register

Structs

R

Register RTSR2 reader

RTSR2_SPEC

rising trigger selection register

W

Register RTSR2 writer

Enums

RT34_A

Rising trigger event configuration bit of Configurable Event input

RT40_A

Rising trigger event configuration bit of Configurable Event input

RT41_A

Rising trigger event configuration bit of Configurable Event input

RT45_A

Rising trigger event configuration bit of Configurable Event input

Type Definitions

RT34_R

Field RT34 reader - Rising trigger event configuration bit of Configurable Event input

RT34_W

Field RT34 writer - Rising trigger event configuration bit of Configurable Event input

RT40_R

Field RT34 reader - Rising trigger event configuration bit of Configurable Event input

RT40_W

Field RT34 writer - Rising trigger event configuration bit of Configurable Event input

RT41_R

Field RT34 reader - Rising trigger event configuration bit of Configurable Event input

RT41_W

Field RT34 writer - Rising trigger event configuration bit of Configurable Event input

RT45_R

Field RT34 reader - Rising trigger event configuration bit of Configurable Event input

RT45_W

Field RT34 writer - Rising trigger event configuration bit of Configurable Event input