Module stm32wlxx_hal::pac::gpioa::afrh

Available on crate feature stm32wl5x_cm4 only.

GPIO alternate function high register

Structs

AFRH_SPEC

GPIO alternate function high register

R

Register AFRH reader

W

Register AFRH writer

Enums

AFRH8_A

Alternate function selection for port x bit y (y = 8..15)

AFRH9_A

Alternate function selection for port x bit y (y = 8..15)

AFRH10_A

Alternate function selection for port x bit y (y = 8..15)

AFRH11_A

Alternate function selection for port x bit y (y = 8..15)

AFRH12_A

Alternate function selection for port x bit y (y = 8..15)

AFRH13_A

Alternate function selection for port x bit y (y = 8..15)

AFRH14_A

Alternate function selection for port x bit y (y = 8..15)

AFRH15_A

Alternate function selection for port x bit y (y = 8..15)

Type Definitions

AFRH8_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH8_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH9_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH9_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH10_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH10_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH11_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH11_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH12_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH12_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH13_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH13_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH14_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH14_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)

AFRH15_R

Field AFRH8 reader - Alternate function selection for port x bit y (y = 8..15)

AFRH15_W

Field AFRH8 writer - Alternate function selection for port x bit y (y = 8..15)