Module stm32wlxx_hal::pac::gpioc::afrl

Available on crate feature stm32wl5x_cm4 only.

GPIO alternate function low register

Structs

AFRL_SPEC

GPIO alternate function low register

R

Register AFRL reader

W

Register AFRL writer

Enums

AFRL0_A

Alternate function selection for port x bit y (y = 0..7)

AFRL1_A

Alternate function selection for port x bit y (y = 0..7)

AFRL2_A

Alternate function selection for port x bit y (y = 0..7)

AFRL3_A

Alternate function selection for port x bit y (y = 0..7)

AFRL4_A

Alternate function selection for port x bit y (y = 0..7)

AFRL5_A

Alternate function selection for port x bit y (y = 0..7)

AFRL6_A

Alternate function selection for port x bit y (y = 0..7)

AFRL7_A

Alternate function selection for port x bit y (y = 0..7)

Type Definitions

AFRL0_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL0_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL1_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL1_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL2_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL2_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL3_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL3_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL4_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL4_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL5_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL5_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL6_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL6_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)

AFRL7_R

Field AFRL0 reader - Alternate function selection for port x bit y (y = 0..7)

AFRL7_W

Field AFRL0 writer - Alternate function selection for port x bit y (y = 0..7)