Module stm32wlxx_hal::pac::pka::sr

Available on crate feature stm32wl5x_cm4 only.

status register

Structs

R

Register SR reader

SR_SPEC

status register

Enums

ADDRERRF_A

Address error flag

BUSY_A

PKA operation is in progressThis bit is set to 1 whenever START bit in the PKA_CR is set. It is automatically cleared when the computation is complete, meaning that PKA RAM can be safely accessed and a new operation can be started.

PROCENDF_A

PKA End of Operation flag

RAMERRF_A

PKA RAM error flag

Type Definitions

ADDRERRF_R

Field ADDRERRF reader - Address error flag

BUSY_R

Field BUSY reader - PKA operation is in progressThis bit is set to 1 whenever START bit in the PKA_CR is set. It is automatically cleared when the computation is complete, meaning that PKA RAM can be safely accessed and a new operation can be started.

PROCENDF_R

Field PROCENDF reader - PKA End of Operation flag

RAMERRF_R

Field RAMERRF reader - PKA RAM error flag