Module stm32wlxx_hal::pac::pwr::cr4

Available on crate feature stm32wl5x_cm4 only.

Power control register 4

Structs

CR4_SPEC

Power control register 4

R

Register CR4 reader

W

Register CR4 writer

Enums

VBE_A

VBAT battery charging enable

VBRS_A

VBAT battery charging resistor selection

WP1_A

Wakeup pin WKUP1 polarity

WP2_A

Wakeup pin WKUP2 polarity

WP3_A

Wakeup pin WKUP3 polarity

WRFBUSYP_A

Wakeup Radio BUSY polarity

Type Definitions

C2BOOT_R

Field C2BOOT reader - oot CPU2 after reset or wakeup from Stop or Standby modes.

C2BOOT_W

Field C2BOOT writer - oot CPU2 after reset or wakeup from Stop or Standby modes.

VBE_R

Field VBE reader - VBAT battery charging enable

VBE_W

Field VBE writer - VBAT battery charging enable

VBRS_R

Field VBRS reader - VBAT battery charging resistor selection

VBRS_W

Field VBRS writer - VBAT battery charging resistor selection

WP1_R

Field WP1 reader - Wakeup pin WKUP1 polarity

WP1_W

Field WP1 writer - Wakeup pin WKUP1 polarity

WP2_R

Field WP2 reader - Wakeup pin WKUP2 polarity

WP2_W

Field WP2 writer - Wakeup pin WKUP2 polarity

WP3_R

Field WP3 reader - Wakeup pin WKUP3 polarity

WP3_W

Field WP3 writer - Wakeup pin WKUP3 polarity

WRFBUSYP_R

Field WRFBUSYP reader - Wakeup Radio BUSY polarity

WRFBUSYP_W

Field WRFBUSYP writer - Wakeup Radio BUSY polarity