Module stm32wlxx_hal::pac::rcc::ahb3smenr

This is supported on crate feature stm32wl5x_cm4 only.

AHB3 peripheral clocks enable in Sleep and Stop modes register

Structs

AESSMEN_R

Field AESSMEN reader - AES accelerator clock enable during CPU1 CSleep mode.

AESSMEN_W

Field AESSMEN writer - AES accelerator clock enable during CPU1 CSleep mode.

AHB3SMENR_SPEC

AHB3 peripheral clocks enable in Sleep and Stop modes register

FLASHSMEN_R

Field FLASHSMEN reader - Flash interface clock enable during CPU1 CSleep mode.

FLASHSMEN_W

Field FLASHSMEN writer - Flash interface clock enable during CPU1 CSleep mode.

PKASMEN_R

Field PKASMEN reader - PKA accelerator clock enable during CPU1 CSleep mode.

PKASMEN_W

Field PKASMEN writer - PKA accelerator clock enable during CPU1 CSleep mode.

R

Register AHB3SMENR reader

RNGSMEN_R

Field RNGSMEN reader - True RNG clocks enable during CPU1 Csleep and CStop modes

RNGSMEN_W

Field RNGSMEN writer - True RNG clocks enable during CPU1 Csleep and CStop modes

SRAM1SMEN_R

Field SRAM1SMEN reader - SRAM1 interface clock enable during CPU1 CSleep mode.

SRAM1SMEN_W

Field SRAM1SMEN writer - SRAM1 interface clock enable during CPU1 CSleep mode.

SRAM2SMEN_R

Field SRAM2SMEN reader - SRAM2 memory interface clock enable during CPU1 CSleep mode

SRAM2SMEN_W

Field SRAM2SMEN writer - SRAM2 memory interface clock enable during CPU1 CSleep mode

W

Register AHB3SMENR writer