Module stm32wlxx_hal::pac::rcc::cr

Available on crate feature stm32wl5x_cm4 only.

Clock control register

Structs

CR_SPEC

Clock control register

R

Register CR reader

W

Register CR writer

Enums

CSSON_A

HSE32 Clock security system enable

HSEBYPPWR_A

Enable HSE32 VDDTCXO output on package pin PB0-VDDTCXO.

HSEON_A

HSE32 clock enable

HSEPRE_A

HSE32 sysclk prescaler

HSERDY_A

HSE32 clock ready flag

HSIASFS_A

HSI16 automatic start from Stop

HSIKERDY_A

HSI16 kernel clock ready flag for peripherals requests.

HSIKERON_A

HSI16 always enable for peripheral kernel clocks.

HSION_A

HSI16 clock enable

HSIRDY_A

HSI16 clock ready flag. (After wakeup from Stop this bit will be read 1 once the HSI16 is ready)

MSION_A

MSI clock enable

MSIPLLEN_A

MSI clock PLL enable

MSIRANGE_A

MSI clock ranges

MSIRDY_A

MSI clock ready flag (After reset this bit will be read 1 once the MSI is ready)

MSIRGSEL_A

MSI range control selection

PLLON_A

Main PLL enable

PLLRDY_A

Main PLL clock ready flag

Type Definitions

CSSON_R

Field CSSON reader - HSE32 Clock security system enable

CSSON_W

Field CSSON writer - HSE32 Clock security system enable

HSEBYPPWR_R

Field HSEBYPPWR reader - Enable HSE32 VDDTCXO output on package pin PB0-VDDTCXO.

HSEBYPPWR_W

Field HSEBYPPWR writer - Enable HSE32 VDDTCXO output on package pin PB0-VDDTCXO.

HSEON_R

Field HSEON reader - HSE32 clock enable

HSEON_W

Field HSEON writer - HSE32 clock enable

HSEPRE_R

Field HSEPRE reader - HSE32 sysclk prescaler

HSEPRE_W

Field HSEPRE writer - HSE32 sysclk prescaler

HSERDY_R

Field HSERDY reader - HSE32 clock ready flag

HSIASFS_R

Field HSIASFS reader - HSI16 automatic start from Stop

HSIASFS_W

Field HSIASFS writer - HSI16 automatic start from Stop

HSIKERDY_R

Field HSIKERDY reader - HSI16 kernel clock ready flag for peripherals requests.

HSIKERON_R

Field HSIKERON reader - HSI16 always enable for peripheral kernel clocks.

HSIKERON_W

Field HSIKERON writer - HSI16 always enable for peripheral kernel clocks.

HSION_R

Field HSION reader - HSI16 clock enable

HSION_W

Field HSION writer - HSI16 clock enable

HSIRDY_R

Field HSIRDY reader - HSI16 clock ready flag. (After wakeup from Stop this bit will be read 1 once the HSI16 is ready)

MSION_R

Field MSION reader - MSI clock enable

MSION_W

Field MSION writer - MSI clock enable

MSIPLLEN_R

Field MSIPLLEN reader - MSI clock PLL enable

MSIPLLEN_W

Field MSIPLLEN writer - MSI clock PLL enable

MSIRANGE_R

Field MSIRANGE reader - MSI clock ranges

MSIRANGE_W

Field MSIRANGE writer - MSI clock ranges

MSIRDY_R

Field MSIRDY reader - MSI clock ready flag (After reset this bit will be read 1 once the MSI is ready)

MSIRGSEL_R

Field MSIRGSEL reader - MSI range control selection

MSIRGSEL_W

Field MSIRGSEL writer - MSI range control selection

PLLON_R

Field PLLON reader - Main PLL enable

PLLON_W

Field PLLON writer - Main PLL enable

PLLRDY_R

Field PLLRDY reader - Main PLL clock ready flag