Module stm32wlxx_hal::subghz

Sub-GHz radio operating in the 150 - 960 MHz ISM band

The main radio type is SubGhz.

LoRa user notice

The Sub-GHz radio may have an undocumented erratum, see this ST community post for more information: link

Structs

BitSync

Bit synchronization.

BpskModParams

BPSK modulation parameters.

BpskPacketParams

Packet parameters for set_lora_packet_params.

CadParams

Channel activity detection (CAD) parameters.

CalibrateImage

Image calibration.

CfgIrq

Argument for set_irq_cfg.

FskBitrate

Bitrate argument for FskModParams::set_bitrate and BpskModParams::set_bitrate.

FskFdev

Frequency deviation argument for FskModParams::set_fdev

FskModParams

(G)FSK modulation parameters.

FskPacketStatus

(G)FSK packet status.

FskStats

Typestate for FSK stats.

GenericPacketParams

Packet parameters for set_packet_params.

HseTrim

HSE32 load capacitor trimming.

LoRaModParams

LoRa modulation parameters.

LoRaPacketParams

Packet parameters for set_lora_packet_params.

LoRaPacketStatus

(G)FSK packet status.

LoRaStats

Typestate for LoRa stats.

PaConfig

Power amplifier configuration parameters.

PktCtrl

Generic packet control.

PwrCtrl

Power control.

RfFreq

RF frequency structure.

SleepCfg

Sleep configuration.

Stats

Packet statistics.

Status

Radio status.

SubGhz

Sub-GHz radio peripheral.

TcxoMode

TCXO trim and HSE32 ready timeout.

Timeout

Timeout argument.

TxParams

Transmit parameters, output power and power amplifier ramp up time.

ValueError

Error for a value that is out-of-bounds.

Enums

AddrComp

Address comparison/filtering for GenericPacketParams.

Calibrate

Block calibration.

CmdStatus

Command status.

CodingRate

LoRa forward error correction coding rate.

CrcType

CRC type definition for GenericPacketParams.

CurrentLim

Power-supply current limit.

ExitMode

Mode to enter after a channel activity detection scan is finished.

FallbackMode

Fallback mode after successful packet transmission or packet reception.

FskBandwidth

Bandwidth options for FskModParams.

FskPulseShape

Pulse shaping options for FskModParams.

HeaderType

Packet header type.

InfSeqSel

Generic packet infinite sequence selection.

Irq

IRQ bit mapping

IrqLine

Interrupt lines.

LoRaBandwidth

LoRa bandwidth.

LoRaSyncWord

LoRa synchronization word.

NbCadSymbol

Number of symbols used for channel activity detection scans.

Ocp

Power amplifier over current protection.

OpError

Operation Errors.

PMode

RX gain power modes.

PaSel

Power amplifier selection.

PacketType

Packet type definition.

PreambleDetection

Preamble detection length for GenericPacketParams.

RampTime

Power amplifier ramp time for FSK, MSK, and LoRa modulation.

RegMode

Radio power supply selection.

RxTimeoutStop

Receiver event which stops the RX timeout timer.

SmpsDrv

SMPS maximum drive capability.

SpreadingFactor

LoRa spreading factor.

StandbyClk

Clock in standby mode.

Startup

Startup configurations when exiting sleep mode.

StatusMode

sub-GHz radio operating mode.

TcxoTrim

TCXO trim.

Functions

mask_irq

Mask the SubGHz IRQ in the NVIC.

rfbusyms

Returns true if the radio is busy or NSS is low.

rfbusys

Returns true if the radio is busy.

unmask_irq⚠

Unmask the SubGHz IRQ in the NVIC.

wakeup⚠

Wakeup the radio from sleep mode.

Type Definitions

Error

Passthrough for SPI errors (for now)