Struct stm32wlxx_hal::subghz::PktCtrl

pub struct PktCtrl { /* fields omitted */ }

Generic packet control.

Argument of set_pkt_ctrl.

Implementations

impl PktCtrl

pub const RESET: PktCtrl

Reset value of the packet control register.

pub const fn from_raw(raw: u8) -> Self

Create a new PktCtrl structure from a raw value.

Reserved bits will be masked.

pub const fn as_bits(&self) -> u8

Get the raw value of the PktCtrl register.

pub const fn set_sync_det_en(self, en: bool) -> PktCtrl

Generic packet synchronization word detection enable.

Example

use stm32wlxx_hal::subghz::PktCtrl;

const PKT_CTRL: PktCtrl = PktCtrl::RESET.set_sync_det_en(true);

pub const fn sync_det_en(&self) -> bool

Returns true if generic packet synchronization word detection is enabled.

Example

use stm32wlxx_hal::subghz::PktCtrl;

let pc: PktCtrl = PktCtrl::RESET;
assert_eq!(pc.sync_det_en(), true);
let pc: PktCtrl = pc.set_sync_det_en(false);
assert_eq!(pc.sync_det_en(), false);
let pc: PktCtrl = pc.set_sync_det_en(true);
assert_eq!(pc.sync_det_en(), true);

pub const fn set_cont_tx_en(self, en: bool) -> PktCtrl

Generic packet continuous transmit enable.

Example

use stm32wlxx_hal::subghz::PktCtrl;

const PKT_CTRL: PktCtrl = PktCtrl::RESET.set_cont_tx_en(true);

pub const fn cont_tx_en(&self) -> bool

Returns true if generic packet continuous transmit is enabled.

Example

use stm32wlxx_hal::subghz::PktCtrl;

let pc: PktCtrl = PktCtrl::RESET;
assert_eq!(pc.cont_tx_en(), false);
let pc: PktCtrl = pc.set_cont_tx_en(true);
assert_eq!(pc.cont_tx_en(), true);
let pc: PktCtrl = pc.set_cont_tx_en(false);
assert_eq!(pc.cont_tx_en(), false);

pub const fn set_inf_seq_sel(self, sel: InfSeqSel) -> PktCtrl

Set the continuous sequence type.

pub const fn inf_seq_sel(&self) -> InfSeqSel

Get the continuous sequence type.

Example

use stm32wlxx_hal::subghz::{InfSeqSel, PktCtrl};

let pc: PktCtrl = PktCtrl::RESET;
assert_eq!(pc.inf_seq_sel(), InfSeqSel::Five);

let pc: PktCtrl = pc.set_inf_seq_sel(InfSeqSel::Zero);
assert_eq!(pc.inf_seq_sel(), InfSeqSel::Zero);

let pc: PktCtrl = pc.set_inf_seq_sel(InfSeqSel::One);
assert_eq!(pc.inf_seq_sel(), InfSeqSel::One);

let pc: PktCtrl = pc.set_inf_seq_sel(InfSeqSel::Prbs9);
assert_eq!(pc.inf_seq_sel(), InfSeqSel::Prbs9);

let pc: PktCtrl = pc.set_inf_seq_sel(InfSeqSel::Five);
assert_eq!(pc.inf_seq_sel(), InfSeqSel::Five);

pub const fn set_inf_seq_en(self, en: bool) -> PktCtrl

Enable infinute sequence generation.

Example

use stm32wlxx_hal::subghz::PktCtrl;

const PKT_CTRL: PktCtrl = PktCtrl::RESET.set_inf_seq_en(true);

pub const fn inf_seq_en(&self) -> bool

Returns true if infinute sequence generation is enabled.

Example

use stm32wlxx_hal::subghz::PktCtrl;

let pc: PktCtrl = PktCtrl::RESET;
assert_eq!(pc.inf_seq_en(), false);
let pc: PktCtrl = pc.set_inf_seq_en(true);
assert_eq!(pc.inf_seq_en(), true);
let pc: PktCtrl = pc.set_inf_seq_en(false);
assert_eq!(pc.inf_seq_en(), false);

pub const fn set_whitening_init(self, val: bool) -> PktCtrl

Set the value of bit-8 (9th bit) for generic packet whitening.

Example

use stm32wlxx_hal::subghz::PktCtrl;

const PKT_CTRL: PktCtrl = PktCtrl::RESET.set_whitening_init(true);

pub const fn whitening_init(&self) -> bool

Returns true if bit-8 of the generic packet whitening is set.

Example

use stm32wlxx_hal::subghz::PktCtrl;

let pc: PktCtrl = PktCtrl::RESET;
assert_eq!(pc.whitening_init(), true);
let pc: PktCtrl = pc.set_whitening_init(false);
assert_eq!(pc.whitening_init(), false);
let pc: PktCtrl = pc.set_whitening_init(true);
assert_eq!(pc.whitening_init(), true);