Struct stm32wlxx_hal::subghz::ValueError

pub struct ValueError<T> { /* private fields */ }

Error for a value that is out-of-bounds.

Used by Timeout::from_duration.

Implementations

impl<T> ValueError<T>

pub const fn too_high(value: T, limit: T) -> ValueError<T>

Create a new ValueError for a value that exceeded an upper bound.

Unfortunately panic is not available in const fn, so there are no guarantees on the value being greater than the limit.

Example

use stm32wlxx_hal::subghz::ValueError;

const ERROR: ValueError<u8> = ValueError::too_high(101u8, 100u8);
assert!(ERROR.over());
assert!(!ERROR.under());

pub const fn too_low(value: T, limit: T) -> ValueError<T>

Create a new ValueError for a value that exceeded a lower bound.

Unfortunately panic is not available in const fn, so there are no guarantees on the value being less than the limit.

Example

use stm32wlxx_hal::subghz::ValueError;

const ERROR: ValueError<u8> = ValueError::too_low(200u8, 201u8);
assert!(ERROR.under());
assert!(!ERROR.over());

pub const fn value(&self) -> &T

Get the value that caused the error.

Example

use stm32wlxx_hal::subghz::ValueError;

const ERROR: ValueError<u8> = ValueError::too_high(101u8, 100u8);
assert_eq!(ERROR.value(), &101u8);

pub const fn limit(&self) -> &T

Get the limit for the value.

Example

use stm32wlxx_hal::subghz::ValueError;

const ERROR: ValueError<u8> = ValueError::too_high(101u8, 100u8);
assert_eq!(ERROR.limit(), &100u8);

pub const fn over(&self) -> bool

Returns true if the value was over the limit.

Example

use stm32wlxx_hal::subghz::ValueError;

const ERROR: ValueError<u8> = ValueError::too_high(101u8, 100u8);
assert!(ERROR.over());
assert!(!ERROR.under());

pub const fn under(&self) -> bool

Returns true if the value was under the limit.

Example

use stm32wlxx_hal::subghz::ValueError;

const ERROR: ValueError<u8> = ValueError::too_low(200u8, 201u8);
assert!(ERROR.under());
assert!(!ERROR.over());