Module stm32wlxx_hal::pac::exti::ftsr1

This is supported on crate feature stm32wl5x_cm4 only.

falling trigger selection register

Structs

FT0_R

Field FT0 reader - Falling trigger event configuration bit of Configurable Event input

FT0_W

Field FT0 writer - Falling trigger event configuration bit of Configurable Event input

FT1_W

Field FT1 writer - Falling trigger event configuration bit of Configurable Event input

FT2_W

Field FT2 writer - Falling trigger event configuration bit of Configurable Event input

FT3_W

Field FT3 writer - Falling trigger event configuration bit of Configurable Event input

FT4_W

Field FT4 writer - Falling trigger event configuration bit of Configurable Event input

FT5_W

Field FT5 writer - Falling trigger event configuration bit of Configurable Event input

FT6_W

Field FT6 writer - Falling trigger event configuration bit of Configurable Event input

FT7_W

Field FT7 writer - Falling trigger event configuration bit of Configurable Event input

FT8_W

Field FT8 writer - Falling trigger event configuration bit of Configurable Event input

FT9_W

Field FT9 writer - Falling trigger event configuration bit of Configurable Event input

FT10_W

Field FT10 writer - Falling trigger event configuration bit of Configurable Event input

FT11_W

Field FT11 writer - Falling trigger event configuration bit of Configurable Event input

FT12_W

Field FT12 writer - Falling trigger event configuration bit of Configurable Event input

FT13_W

Field FT13 writer - Falling trigger event configuration bit of Configurable Event input

FT14_W

Field FT14 writer - Falling trigger event configuration bit of Configurable Event input

FT15_W

Field FT15 writer - Falling trigger event configuration bit of Configurable Event input

FT16_W

Field FT16 writer - Falling trigger event configuration bit of Configurable Event input

FT21_W

Field FT21 writer - Falling trigger event configuration bit of Configurable Event input

FT22_W

Field FT22 writer - Falling trigger event configuration bit of Configurable Event input

FTSR1_SPEC

falling trigger selection register

R

Register FTSR1 reader

W

Register FTSR1 writer

Enums

FT0_A

Falling trigger event configuration bit of Configurable Event input

Type Definitions

FT1_A

Falling trigger event configuration bit of Configurable Event input

FT1_R

Field FT1 reader - Falling trigger event configuration bit of Configurable Event input

FT2_A

Falling trigger event configuration bit of Configurable Event input

FT2_R

Field FT2 reader - Falling trigger event configuration bit of Configurable Event input

FT3_A

Falling trigger event configuration bit of Configurable Event input

FT3_R

Field FT3 reader - Falling trigger event configuration bit of Configurable Event input

FT4_A

Falling trigger event configuration bit of Configurable Event input

FT4_R

Field FT4 reader - Falling trigger event configuration bit of Configurable Event input

FT5_A

Falling trigger event configuration bit of Configurable Event input

FT5_R

Field FT5 reader - Falling trigger event configuration bit of Configurable Event input

FT6_A

Falling trigger event configuration bit of Configurable Event input

FT6_R

Field FT6 reader - Falling trigger event configuration bit of Configurable Event input

FT7_A

Falling trigger event configuration bit of Configurable Event input

FT7_R

Field FT7 reader - Falling trigger event configuration bit of Configurable Event input

FT8_A

Falling trigger event configuration bit of Configurable Event input

FT8_R

Field FT8 reader - Falling trigger event configuration bit of Configurable Event input

FT9_A

Falling trigger event configuration bit of Configurable Event input

FT9_R

Field FT9 reader - Falling trigger event configuration bit of Configurable Event input

FT10_A

Falling trigger event configuration bit of Configurable Event input

FT10_R

Field FT10 reader - Falling trigger event configuration bit of Configurable Event input

FT11_A

Falling trigger event configuration bit of Configurable Event input

FT11_R

Field FT11 reader - Falling trigger event configuration bit of Configurable Event input

FT12_A

Falling trigger event configuration bit of Configurable Event input

FT12_R

Field FT12 reader - Falling trigger event configuration bit of Configurable Event input

FT13_A

Falling trigger event configuration bit of Configurable Event input

FT13_R

Field FT13 reader - Falling trigger event configuration bit of Configurable Event input

FT14_A

Falling trigger event configuration bit of Configurable Event input

FT14_R

Field FT14 reader - Falling trigger event configuration bit of Configurable Event input

FT15_A

Falling trigger event configuration bit of Configurable Event input

FT15_R

Field FT15 reader - Falling trigger event configuration bit of Configurable Event input

FT16_A

Falling trigger event configuration bit of Configurable Event input

FT16_R

Field FT16 reader - Falling trigger event configuration bit of Configurable Event input

FT21_A

Falling trigger event configuration bit of Configurable Event input

FT21_R

Field FT21 reader - Falling trigger event configuration bit of Configurable Event input

FT22_A

Falling trigger event configuration bit of Configurable Event input

FT22_R

Field FT22 reader - Falling trigger event configuration bit of Configurable Event input