Module stm32wlxx_hal::pac::exti::rtsr2

This is supported on crate feature stm32wl5x_cm4 only.

rising trigger selection register

Structs

R

Register RTSR2 reader

RT34_R

Field RT34 reader - Rising trigger event configuration bit of Configurable Event input

RT34_W

Field RT34 writer - Rising trigger event configuration bit of Configurable Event input

RT40_W

Field RT40 writer - Rising trigger event configuration bit of Configurable Event input

RT41_W

Field RT41 writer - Rising trigger event configuration bit of Configurable Event input

RT45_W

Field RT45 writer - Rising trigger event configuration bit of Configurable Event input

RTSR2_SPEC

rising trigger selection register

W

Register RTSR2 writer

Enums

RT34_A

Rising trigger event configuration bit of Configurable Event input

Type Definitions

RT40_A

Rising trigger event configuration bit of Configurable Event input

RT40_R

Field RT40 reader - Rising trigger event configuration bit of Configurable Event input

RT41_A

Rising trigger event configuration bit of Configurable Event input

RT41_R

Field RT41 reader - Rising trigger event configuration bit of Configurable Event input

RT45_A

Rising trigger event configuration bit of Configurable Event input

RT45_R

Field RT45 reader - Rising trigger event configuration bit of Configurable Event input