Function stm32wlxx_hal::pwr::setup_wakeup_pins

pub fn setup_wakeup_pins(
    pwr: &mut PWR, 
    wp1: WakeupPin, 
    wp2: WakeupPin, 
    wp3: WakeupPin
)

Setup the wakeup pins for shutdown and standby low-power modes.

The reference manual is not 100% specific which wakeup pin corresponds to which physical pin, a footnote of “Table 45. Functionalities depending on system operating mode” in RM0453 Rev 2 implies the following:

WP1 corresponds to A0

WP2 corresponds to C13

WP3 corresponds to B3

If you know where to find more concrete information on this please open an issue.

Example

Enable A0 to wakeup on a falling edge.

use stm32wlxx_hal::{
    pac,
    pwr::{setup_wakeup_pins, WakeupPin},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();
setup_wakeup_pins(
    &mut dp.PWR,
    WakeupPin::Falling,
    WakeupPin::Disabled,
    WakeupPin::Disabled,
);