Module stm32wlxx_hal::rcc

Reset and clocking control

Unlike other HALs clocks do not get frozen. A lot of applications for this chip will require low-power, and there are many scenarios where you will want to adjust the clocks.

Quickstart: set_sysclk_msi_max

Enums

LsiPre

LSI frequency prescaler

MsiRange

MSI clock ranges.

Vos

Voltage scaling

Functions

cpu1_systick_hz

Calculate the current CPU1 systick frequency in hertz

cpu2_systick_hzstm32wl5x_cm4 or stm32wl5x_cm0p

Calculate the current CPU2 systick frequency in hertz

cpu_systick_hz

Calculate the current CPU systick frequency in hertz.

enable_lsi

Enable the LSI clock with the currently configured pre-scaler and wait for completion.

hclk1_hz

Calculate the current hclk1 frequency in hertz

hclk2_hzstm32wl5x_cm4 or stm32wl5x_cm0p

Calculate the current hclk2 frequency in hertz

hclk3_hz

Calculate the current hclk3 frequency in hertz

lsi_hz

Calculate the LSI clock frequency in hertz.

pclk1_hz

Calculate the current PCLK1 frequency in hertz

pclk2_hz

Calculate the current PCLK2 frequency in hertz

pulse_reset_backup_domain⚠

Reset the backup domain.

set_sysclk_hse⚠

Set the sysclk to use the HSE 32MHz clock.

set_sysclk_hsi⚠

Set the sysclk to use the HSI 16MHz clock.

set_sysclk_msi⚠

Set the sysclk from an MSI range.

set_sysclk_msi_max⚠

Set the sysclk to the MSI source at 48MHz.

setup_lsi⚠

Setup the LSI clock and wait for completion.

sysclk_hz

Calculate the current system clock frequency in hertz