Struct stm32wlxx_hal::rtc::Rtc

pub struct Rtc { /* fields omitted */ }

Real-time clock driver.

Implementations

impl Rtc

pub fn new(rtc: RTC, clk: Clk, pwr: &mut PWR, rcc: &mut RCC) -> Rtc

Create a new real-time clock driver.

This will not setup the source clock.

Safety

This function could be considered unsafe because it is not a pure function. The RTC is in the backup domain; system resets will not reset the RTC. You are responsible for resetting the backup domain if required.

Panics

(debug) clock source is not ready.

Example

LSE clock source (this depends on HW, example valid for NUCLEO board):

use stm32wlxx_hal::{
    pac,
    rcc::pulse_reset_backup_domain,
    rtc::{Clk, Rtc},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

unsafe { pulse_reset_backup_domain(&mut dp.RCC, &mut dp.PWR) };
dp.PWR.cr1.modify(|_, w| w.dbp().enabled());
dp.RCC.bdcr.modify(|_, w| w.lseon().on());
while dp.RCC.bdcr.read().lserdy().is_not_ready() {}

let rtc: Rtc = Rtc::new(dp.RTC, Clk::Lse, &mut dp.PWR, &mut dp.RCC);

LSI clock source:

use stm32wlxx_hal::{
    pac,
    rcc::{enable_lsi, pulse_reset_backup_domain},
    rtc::{Clk, Rtc},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

unsafe { pulse_reset_backup_domain(&mut dp.RCC, &mut dp.PWR) };
enable_lsi(&mut dp.RCC);

let rtc: Rtc = Rtc::new(dp.RTC, Clk::Lsi, &mut dp.PWR, &mut dp.RCC);

HSE clock source (this depends on HW, example valid for NUCLEO board):

use stm32wlxx_hal::{
    pac,
    rcc::pulse_reset_backup_domain,
    rtc::{Clk, Rtc},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

unsafe { pulse_reset_backup_domain(&mut dp.RCC, &mut dp.PWR) };
dp.RCC
    .cr
    .modify(|_, w| w.hseon().enabled().hsebyppwr().vddtcxo());
while dp.RCC.cr.read().hserdy().is_not_ready() {}

let rtc: Rtc = Rtc::new(dp.RTC, Clk::Hse, &mut dp.PWR, &mut dp.RCC);

pub unsafe fn renew(rtc: RTC, pwr: &mut PWR, rcc: &mut RCC) -> Rtc

Create a new real-time clock driver preserving backup domain values.

Unlike new this will enable the LSE clock source if not already enabled. This function assumes the LSE clock source will be used because it is the only clock source that is preserved in the shutdown mode.

The RTC calendar may not be initialized, this can occur if this function is called after power loss or after a backup domain reset.

Safety

This function relies on global hardware state in the backup domain. The backup domain is not reset with normal system resets. Reset the backup domain before calling this function if determinism is required.

pub fn hz(rcc: &RCC) -> u32

Source clock frequency in hertz.

pub fn status() -> R

Read the RTC status (interrupt) register.

pub fn masked_status() -> R

Read the RTC masked status (interrupt) register.

pub fn clear_status(mask: u32)

Clear status (interrupt) flags.

Status flag masks can be found in stat.

pub fn set_date_time(&mut self, date_time: NaiveDateTime)

Set the date and time.

The value will take some duration to apply after this function returns:

LPCAL=0: the counting restarts after 4 RTCCLK clock cycles

LPCAL=1: the counting restarts after up to 2 RTCCLK + 1 ck_apre

Panics

Year is greater than or equal to 2100.

Year is less than 2000.

Backup domain write protection is enabled.

pub fn calendar_initialized(&self) -> Option<()>

Returns None if the calendar is uninitialized.

pub fn date(&self) -> Option<NaiveDate>

Calendar Date

Returns None if the calendar has not been initialized.

pub fn time(&self) -> Option<NaiveTime>

Current Time

Returns None if the calendar has not been initialized.

pub fn date_time(&self) -> Option<NaiveDateTime>

Calendar Date and Time

Returns None if the calendar has not been initialized.

pub fn setup_wakeup_timer(&mut self, sec: u32, irq_en: bool)

Setup the periodic wakeup timer for sec + 1 seconds.

sec can only go up to 217 (36 hours), values greater than this will be set to the maximum.

Example

Setup the wakeup timer to go off in 1 hour, without interrupts.

rtc.setup_wakeup_timer(3599, false);

pub fn disable_wakeup_timer(&mut self)

Disable the wakeup timer.

pub fn disable_write_protect(&mut self)

Disable the RTC write protection.

pub unsafe fn enable_write_protect(&mut self)

Enable the RTC write protection.

Safety

You must call disable_write_protect before using any other &mut self RTC method.