Struct stm32wlxx_hal::uart::Uart1

pub struct Uart1<RX, TX> { /* fields omitted */ }

UART1 driver.

Implementations

impl Uart1<NoRx, NoTx>

pub fn new(
uart: USART1,
baud: u32,
clk: Clk,
rcc: &mut RCC
) -> Uart1<NoRx, NoTx>

Create a new UART driver from a UART peripheral.

This will enable clocks and reset the UART peripheral.

Example

use stm32wlxx_hal::{
    pac,
    uart::{self, NoRx, NoTx, Uart1},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let uart: Uart1<NoRx, NoTx> = Uart1::new(dp.USART1, 115_200, uart::Clk::Hsi16, &mut dp.RCC);

impl<RX, TX> Uart1<RX, TX>

pub fn clock_hz(&self, rcc: &RCC) -> u32

Calculate the clock frequency.

Fractional frequencies will be rounded towards zero.

impl Uart1<NoRx, NoTx>

pub unsafe fn pulse_reset(rcc: &mut RCC)

Reset the UART.

Safety

The UART must not be in-use.

You are reponsible for setting up the UART after a reset.

Example

See steal

impl Uart1<NoRx, NoTx>

pub fn enable_clock(rcc: &mut RCC)

Enable the UART clock.

This is done for you in new

Example

use stm32wlxx_hal::{pac, uart::LpUart};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();
LpUart::enable_clock(&mut dp.RCC);

pub unsafe fn disable_clock(rcc: &mut RCC)

Disable the UART clock.

Safety

You are responsible for ensuring the UART is in a state where the clock can be disabled without entering an error state.

You cannot use the UART while the clock is disabled.

You are responsible for re-enabling the clock before resuming use of the UART.

You are reponsible for setting up anything that may have lost state while the clock was disabled.

impl Uart1<NoRx, NoTx>

pub unsafe fn steal() -> Uart1<NoRx, NoTx>

Steal the UART peripheral from whatever is currently using it.

This will not initialize the peripheral (unlike new).

Safety

Ensure that the code stealing the UART has exclusive access to the peripheral. Singleton checks are bypassed with this method.

You are responsible for setting up the UART correctly.

Example

use stm32wlxx_hal::{
    pac,
    uart::{LpUart, NoRx, NoTx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

LpUart::enable_clock(&mut dp.RCC);
// safety:
// 1. Nothing else is using the LPUART in this code.
// 2. This code performs setup for the LPUART.
unsafe { LpUart::pulse_reset(&mut dp.RCC) };

// safety:
// 1. Nothing else is using the LPUART in this code.
// 2. The LPUART has been setup, clocks are enabled and the LPUART has been reset.
let mut lpuart: LpUart<NoRx, NoTx> = unsafe { LpUart::steal() };

impl<RX, TX> Uart1<RX, TX>

pub fn free(self) -> USART1

Free the UART peripheral from the driver.

Example

use stm32wlxx_hal::{
    pac,
    uart::{self, LpUart, NoRx, NoTx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();
let lpuart: pac::LPUART = dp.LPUART;

let lpuart: LpUart<NoRx, NoTx> =
    LpUart::new(lpuart, 115_200, uart::Clk::Hsi16, &mut dp.RCC);
// ... use LPUART
let lpuart: pac::LPUART = lpuart.free();

impl<RX> Uart1<RX, NoTx>

pub fn enable_tx<TX: Uart1Tx>(
self,
tx: TX,
cs: &CriticalSection
) -> Uart1<RX, TX>

Enable the UART transmitter.

Example

use stm32wlxx_hal::{
    cortex_m,
    gpio::{pins, PortB},
    pac,
    uart::{self, LpUart, NoRx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let uart: LpUart<NoRx, pins::B11> = cortex_m::interrupt::free(|cs| {
    LpUart::new(dp.LPUART, 115_200, uart::Clk::Hsi16, &mut dp.RCC)
        .enable_tx(gpiob.b11, cs)
});

impl<RX> Uart1<RX, NoTx>

pub fn enable_tx_dma<TxPin: Uart1Tx, TxDma: DmaCh>(
self,
tx: TxPin,
tx_dma: TxDma,
cs: &CriticalSection
) -> Uart1<RX, (TxPin, TxDma)>

Enable the UART transmitter with a DMA channel.

Example

use stm32wlxx_hal::{
    cortex_m,
    dma::{AllDma, Dma2Ch7},
    gpio::{pins, PortB},
    pac,
    uart::{self, LpUart, NoRx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let dma: AllDma = AllDma::split(dp.DMAMUX, dp.DMA1, dp.DMA2, &mut dp.RCC);
let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let uart: LpUart<NoRx, (pins::B11, Dma2Ch7)> = cortex_m::interrupt::free(|cs| {
    LpUart::new(dp.LPUART, 115_200, uart::Clk::Hsi16, &mut dp.RCC)
        .enable_tx_dma(gpiob.b11, dma.d2.c7, cs)
});

impl<RX, TX> Uart1<RX, TX>

pub fn disable_tx(self) -> (Uart1<RX, NoTx>, TX)

Disable the UART transmitter.

Example

use stm32wlxx_hal::{
    cortex_m,
    gpio::{pins, PortB},
    pac,
    uart::{self, LpUart, NoRx, NoTx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let uart: LpUart<NoRx, pins::B11> = cortex_m::interrupt::free(|cs| {
    LpUart::new(dp.LPUART, 115_200, uart::Clk::Hsi16, &mut dp.RCC)
        .enable_tx(gpiob.b11, cs)
});

let (uart, b11): (LpUart<NoRx, NoTx>, pins::B11) = uart.disable_tx();

impl<TX> Uart1<NoRx, TX>

pub fn enable_rx<RX: Uart1Rx>(
self,
rx: RX,
cs: &CriticalSection
) -> Uart1<RX, TX>

Enable the UART receiver.

Example

use stm32wlxx_hal::{
    cortex_m,
    gpio::{pins, PortB},
    pac,
    uart::{self, LpUart, NoTx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let uart: LpUart<pins::B10, NoTx> = cortex_m::interrupt::free(|cs| {
    LpUart::new(dp.LPUART, 115_200, uart::Clk::Hsi16, &mut dp.RCC)
        .enable_rx(gpiob.b10, cs)
});

impl<TX> Uart1<NoRx, TX>

pub fn enable_rx_dma<RxPin: Uart1Rx, RxDma: DmaCh>(
self,
rx: RxPin,
rx_dma: RxDma,
cs: &CriticalSection
) -> Uart1<(RxPin, RxDma), TX>

Enable the UART receiver with a DMA channel.

Example

use stm32wlxx_hal::{
    dma::{AllDma, Dma2Ch2},
    gpio::{pins, PortB},
    pac,
    uart::{self, LpUart, NoTx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let dma: AllDma = AllDma::split(dp.DMAMUX, dp.DMA1, dp.DMA2, &mut dp.RCC);
let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let uart: LpUart<(pins::B10, Dma2Ch2), NoTx> = cortex_m::interrupt::free(|cs| {
    LpUart::new(dp.LPUART, 115_200, uart::Clk::Hsi16, &mut dp.RCC)
        .enable_rx_dma(gpiob.b10, dma.d2.c2, cs)
});

impl<RX, TX> Uart1<RX, TX>

pub fn disable_rx(self) -> (Uart1<NoRx, TX>, RX)

Disable the UART receiver.

Example

use stm32wlxx_hal::{
    gpio::{pins, PortB},
    pac,
    uart::{self, LpUart, NoRx, NoTx},
};

let mut dp: pac::Peripherals = pac::Peripherals::take().unwrap();

// enable the HSI16 source clock
dp.RCC.cr.modify(|_, w| w.hsion().set_bit());
while dp.RCC.cr.read().hsirdy().is_not_ready() {}

let gpiob: PortB = PortB::split(dp.GPIOB, &mut dp.RCC);
let uart: LpUart<pins::B10, NoTx> = cortex_m::interrupt::free(|cs| {
    LpUart::new(dp.LPUART, 115_200, uart::Clk::Hsi16, &mut dp.RCC)
        .enable_rx(gpiob.b10, cs)
});

let (uart, b10): (LpUart<NoRx, NoTx>, pins::B10) = uart.disable_rx();

impl<RxPin, RxDma, TX> Uart1<(RxPin, RxDma), TX> where
RxPin: Uart1Rx,
RxDma: DmaCh,

pub fn bread_all(&mut self, buffer: &mut [u8]) -> Result<(), Error>

This is not an embedded-hal trait, it is added simply for parity with what exists on the TX side.